moved template files to S4 default dir
