+++ src/DataFixtures/CommentFixtures.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace App\DataFixtures\ORM;
+namespace App\DataFixtures;
 
 use App\Entity\Article;
-use Doctrine\Common\DataFixtures\AbstractFixture;
 use Doctrine\Common\DataFixtures\DependentFixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use App\Entity\Comment;
 use Faker\Factory;
+use Doctrine\Bundle\FixturesBundle\Fixture;
 
-class CommentFixtures extends AbstractFixture implements DependentFixtureInterface
+class CommentFixtures extends Fixture implements DependentFixtureInterface
 {
     public function load(ObjectManager $manager)
     {
