--- src/Resources/config/services.yml
@@ -1,45 +1,0 @@
-services:
-
-    form.type.enquire_type:
-        class: App\Form\EnquiryType
-        tags: [form.choice_list_factor]
-
-    blog.fixtures.article:
-        class: App\DataFixtures\ORM\ArticleFixtures
-        tags: [doctrine.fixture.orm]
-
-    blog.fixtures.comment:
-        class: App\DataFixtures\ORM\CommentFixtures
-        tags: [doctrine.fixture.orm]
-
-    blog.twig.extension:
-        class: App\Twig\Extensions\BlogExtension
-        tags: [twig.extension]
-
-    pagerfanta.view.default:
-        class: Pagerfanta\View\DefaultView
-        public: false
-        tags: [{ name: pagerfanta.view, alias: default }]
-
-    phpro.captcha-service:
-        class: App\Utils\Recaptcha
-        arguments: ["%recaptcha_sitekey%", '@logger']
-
-    custom.event.contact_page:
-        class: App\EventSubscriber\ContactPageEventSubscriber
-        tags:
-            - { name: kernel.event_listener, event: custom.event.contact_page, method: onCustomEvent }
-        arguments: ['@phpro.mailer.mailer_service']
-
-    phpro.mailer.mailer_service:
-        class: App\Mailer\MailerService
-        arguments:
-            - '@twig'
-            - '@swiftmailer.mailer'
-        calls:
-            - method: setLogger
-              arguments:
-                  - '@logger'
-
-    phpro.extractor.tag_extractor:
-        class: App\Extractor\TagExtractor
