--- src/Resources/views/base.html.twig
@@ -1,58 +1,0 @@
-<!-- app/Resources/views/base.html.twig -->
-<!DOCTYPE html>
-<html>
-<head>
-    <meta http-equiv="Content-Type" content="text/html"; charset=utf-8" />
-    <title>{% block title %}symblog{% endblock %} - symblog</title>
-    <!--[if lt IE 9]>
-    <script src="http://html5shim.googlecode.com/svn/trunk/html5.js"></script>
-    <![endif]-->
-    {% block stylesheets %}
-        <link href='http://fonts.googleapis.com/css?family=Irish+Grover' rel='stylesheet' type='text/css'>
-        <link href='http://fonts.googleapis.com/css?family=La+Belle+Aurore' rel='stylesheet' type='text/css'>
-        <link href="../css/screen.css" type="text/css" rel="stylesheet" />
-        <link rel="stylesheet" href="{{ asset('bundles/whiteoctoberpagerfanta/css/pagerfantaDefault.css') }}" type="text/css" media="all" />
-    {% endblock %}
-</head>
-<body>
-
-<section id="wrapper">
-    <header id="header">
-        <div class="top">
-            {% block navigation %}
-                <nav>
-                    <ul class="navigation">
-                        <li><a href="{{ path('blog_homepage') }}">{% trans %}blog.home_nav_title{% endtrans %}</a></li>
-                        <li><a href="{{ path('page_about') }}">{% trans %}blog.about_nav_title{% endtrans %}</a></li>
-                        <li><a href="{{ path('page_contact') }}">{% trans %}blog.contact_nav_title{% endtrans %}</a></li>
-                    </ul>
-                </nav>
-            {% endblock %}
-        </div>
-
-        <hgroup>
-            <h2>{% block blog_title %}<a href="{{ path('blog_homepage') }}">symblog</a>{% endblock %}</h2>
-            <h3>{% block blog_tagline %}<a href="{{ path('blog_homepage') }}">{% trans %}blog.title{% endtrans %}</a>{% endblock %}</h3>
-        </hgroup>
-    </header>
-
-    <section class="main-col">
-        {% block body %}{% endblock %}
-    </section>
-    <aside class="sidebar">
-        {% block sidebar %}{% endblock %}
-    </aside>
-
-    <div id="footer">
-        {% block footer %}
-            Symfony2 blog tutorial - {% trans %}blog.created.by{% endtrans %} <a href="https://github.com/dsyph3r">dsyph3r</a>
-        {% endblock %}
-    </div>
-</section>
-
-{% block javascripts %}
-    <script src="https://www.google.com/recaptcha/api.js" async defer></script>
-    <script src="../js/reCaptcha.js"></script>
-{% endblock %}
-</body>
-</html>
