--- src/Resources/views/layout.html.twig
@@ -1,11 +1,0 @@
-{% extends 'base.html.twig' %}
-{% block stylesheets %}
-    {{ parent() }}
-    <link href="../css/blog.css" type="text/css" rel="stylesheet" />
-    <link href="../css/sidebar.css" type="text/css" rel="stylesheet" />
-{% endblock %}
-
-
-{% block sidebar %}
-    {{ render(controller("App:Page:sidebar", { 'call' : 'action' })) }}
-{% endblock %}
