+++ templates/Page/sidebar.html.twig
@@ -4,7 +4,7 @@
     </header>
     <p class="tags">
         {% for category in categories %}
-        <a href="{{ path('blog_show_by_categorie', { 'category':  category}) }}">
+        <a href="{{ path('blog_category_show', { 'category':  category}) }}">
             {{ category }}
         </a>
         {% else %}
